+++ aleth/main.cpp
@@ -321,8 +321,9 @@ int main(int argc, char** argv)
     addImportExportOption("import-snapshot", po::value<string>()->value_name("<path>"),
         "Import blockchain and state data from the Parity Warp Sync snapshot\n");
 
-    ///List_logging_channels
-    std::string const logChannels = "\nblock blockhdr bq chain client debug discov error ethcap exec host impolite info net overlaydb peer rpc snap statedb sync timer tq trace vmtrace warn watch";
+    std::string const logChannels =
+        "block blockhdr bq chain client debug discov error ethcap exec host impolite info net "
+        "overlaydb peer rlpx rpc snap statedb sync timer tq trace vmtrace warn watch";
     LoggingOptions loggingOptions;
     po::options_description loggingProgramOptions(
         createLoggingProgramOptions(c_lineWidth, loggingOptions, logChannels));